backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
yfinance==0.2.32
pandas==2.1.3
python-dotenv==1.0.0
httpx==0.27.0
pydantic==2.5.1
python-multipart==0.0.6
supabase==2.13.0
groq
